Moved swaggerDefinition to docs folder
